requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.5.0
pymongo==4.2.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
python-engineio==4.3.3
python-socketio==5.7.1
simple-websocket==0.7.0
Flask-Login==0.6.2
Flask-SocketIO==5.2.0  